Simplified scripts and created a new function. QC measures were adjusted
Most important changes:
Adjusted quality metrics.
Defined new function to map things to a specific column (1-0) called map_to_column.
Replaced the mapping lines with the new function.
Modified the filtering function to allow for skipping n_genes_by_count when entering a value of 0.
Created another line of filtering by separating patient antrum and writing to disk. Must check if combind_patient_epithelium is correct or if I'm looking to just spearate the antrum.
Changes had to happen when the written files that were an output for volcano. Need to check that later but will commit for now.